FIX src posix -- RWX memory for code space
In new versions of Linux, depending on the BIOS setup and system
configuration, the memory regions allocated for code must be explicitly
configured as executable.
For example, see https://wiki.ubuntu.com/Security/Features#nx

For this reason, a new allocation method, `ALLOCATE-RWX`, has been
introduced. This method is now used to allocate memory for dynamically
allocated dictionaries and temporary wordlists.

Thanks to @ivanivanovo for reporting of the initial problem
and research for possible solutions: https://t.me/spforthchat/764