FFA src compiler -- migration to name tokens API
Now the name token for a word is represented by the word's NFA field.
However, NFA is not a part of the system API, but rather an internal
implementation detail. The user must not assume that nt is NFA
(or a counted string address). So, special access methods must be used
to access other details of the word from its nt.

In word names, nt, as the successor of NFA, is denoted as "NAME" (by tradition).
For example, "NAME> ( nt -- xt )"

The following words have been added:
- "NAME>NEXT-NAME" instead of "CDR" -- a longer name, less clashes
  - the old name is kept as a synonym
- "NAME>STRING" (it's moved into core from lib/include/wordlist-tools.f)
- "NAME>CSTRING" (it's useful in many cases, and it hides information about
    the structure of nt)
- "LATEST-NAME" and "LATEST-NAME-IN" (as its factor)
- "LATEST-NAME-XT"

The old word "LATEST" may return nt for the current definition, and may return 0.
The new word "LATEST-NAME" returns only nt for the latest definition in the
compilation word list, which is also available via "TRAVERSE-WORDLIST",
or throws an exception.

Now "IMMEDIATE" and "VOC" affect the definition that is returned by
"LATEST-NAME", instead of "LAST @".

A bug in Locals re IMMEDIATE is fixed. Now the following test is passed:
  t{ : foo { a } ; immediate s" foo" sfind nip -> 1 }t

A bug in TRAVERSE-WORDLIST is fixed: the current definition was available via
this word (see "SMUDGE").

The use of "{wid} @", "LAST @", "LAST-CFA @", and sometimes "LATEST"
are replaced by the new words in "src/" and "lib/".
Additionally, to set a word list name, the setter "VOC-NAME!" is used,
insetead of direct manipulation of addresses.