FAC devel -- refactoring, migration to name tokens API
Now NFA is not a part of the system API, but rather an internal
implementation detail. The user must not assume that nt is NFA
(or a counted string address). So, special access methods must be used
to access other details of the word from its nt.

The old word "LATEST" may return nt for the current definition, and may return 0.
The new word "LATEST-NAME" returns only nt for the latest definition in the
compilation word list, which is also available via "TRAVERSE-WORDLIST",
or throws an exception.

The use of "{wid} @", "LAST @", "LAST-CFA @", and sometimes "LATEST"
are replaced in "devel/" by "LATEST-NAME", when appropriate.

Additionally, to set a word list name, the setter "VOC-NAME!" is used,
insetead of direct manipulation of addresses.