DEL lib include ansi-current -- it is against Forth-94
The module "lib/include/ansi-current.f" just allows
a program to violate the Forth-94 (ANSI) and Forth-2012 standards,
by manage some side effects when the compilation word list is changed
before the current definition is ended.

See 16.3.3 Finding definition names
An ambiguous condition exists if a program changes the compilation word list
during the compilation of a definition or before modification of the behavior
of the most recently compiled definition with `;CODE`, `DOES>`, or `IMMEDIATE`.