FIX lib asm -- incorrect use of "IMMEDIATE"
The definition for ";M" (via "MACRO: ;M") was made immediate
when the compilation word list was already changed. It's incorrect.
This bug is fixed via introducing "ENDM-IMM", which ends the macro
definition and also makes it immediate.

Along with this bug fix, the synonym "FORTH.IMMEDIATE" for "IMMEDIATE"
was introduced in the vocabulary ASM-HIDDEN to make some code shorter.