FAC src tc -- eliminate code duplication
Move definitions for words that are missing in the host Forth system
to "src/spf.f", explicitly declare spf4-specific definitions.

In some cases the intention was to use `EXECUTE` from `TC`, but
the host's `EXECUTE` is used instead, since `EXECUTE` is defined below,
and not in `TC` but in `TC-IMM`. The corresponding unnecessary exercises
with the search order about `TC` have now been removed.